Adicionei os Desafios 1 e 2 ao exercício

+++ modulo2/apis-labenusers/apis-labenusers-app/src/componentes/ListaDeUsuarios.js
@@ -1,10 +1,15 @@
 import React from "react"
 import axios from "axios"
+import DetalhesUsuario from "./DetalhesUsuario"
 
 class ListaDeUsuarios extends React.Component {
     state = {
         valorInputSearch: "",
-        usuarios: []
+        usuarios: [],
+        detalhes: false,
+        nameDetalhes: "",
+        emailDetalhes: "",
+        idDetalhes: ""
     }
 
     getAllUsers = () => {
@@ -25,19 +30,57 @@ class ListaDeUsuarios extends React.Component {
     }
 
     deleteUser = (event) => {
+        if(window.confirm(`Tem certeza que deseja deletar este usuário?`)){
+            const id = event.target.id
+    
+            const requestExcluir = axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, {
+                headers: {
+                    Authorization: "rafael-correia-freire"
+                }
+            })
+    
+            requestExcluir.then(() => {
+                alert('Usuário excluído com sucesso')
+            }).catch(error => {
+                alert(error.message)
+            })
+
+            this.setState({
+                nameDetalhes: "",
+                emailDetalhes: "",
+                detalhes: false
+            })
+        }
+    }
+
+    handleDetalhes = (event) => {
         const id = event.target.id
+        if(this.state.detalhes) {
+            this.setState({
+                nameDetalhes: "",
+                emailDetalhes: "",
+                detalhes: false
+            })
+        }
+        else {
+            const getUserById = axios.get(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, {
+                headers: {
+                    Authorization: "rafael-correia-freire"
+                }
+            })
 
-        const requestExcluir = axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, {
-            headers: {
-                Authorization: "rafael-correia-freire"
-            }
-        })
-
-        requestExcluir.then(() => {
-            alert('Usuário excluído com sucesso')
-        }).catch(error => {
-            alert(error.message)
-        })
+            getUserById.then((response) => {
+                this.setState({
+                    nameDetalhes: "Nome: " + response.data.name,
+                    emailDetalhes: "Email: " + response.data.email,
+                    detalhes: true,
+                    idDetalhes: response.data.id
+                })
+            })
+            .catch(error => {
+                alert(error.message)
+            })
+        }
     }
 
     componentDidMount = () => {
@@ -48,13 +91,12 @@ class ListaDeUsuarios extends React.Component {
         this.getAllUsers()
     }
 
-
     render() {
         const listaUsuarios = this.state.usuarios.map(usuario => {
             return (
-                <li key={usuario.id}>
-                    {usuario.name}
-                    <button type="button" onClick={this.deleteUser} id={usuario.id} >❌</button>
+                <li key={usuario.id} onClick={this.handleDetalhes} id={usuario.id}>
+                    <span>{usuario.name}</span>
+                    <button type="button" onClick={this.deleteUser} id={usuario.id}>❌</button>
                 </li>
             )
         })
@@ -65,6 +107,13 @@ class ListaDeUsuarios extends React.Component {
                 <ul>
                     {listaUsuarios}
                 </ul>
+                <DetalhesUsuario
+                    name={this.state.nameDetalhes}
+                    email={this.state.emailDetalhes}
+                    onClick={this.deleteUser}
+                    detalhes={this.state.detalhes}
+                    id={this.state.idDetalhes}
+                />
             </div>
         )
     }
